+++ src/pages/Login/LoginPage/LoginPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import * as S from "./Styles";
 import Leave from "../../../assets/icons/Leave.svg";
-import Button from "../../../components/Login/KakaoButton";
+import Button from "../_components/KakaoButton";
 
 const LoginPage: React.FC = () => {
     return (
